snackbar部分変更

+++ app/src/main/java/jp/techacademy/yutaro/calcapp/SecondActivity.kt
@@ -12,7 +12,7 @@ class SecondActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_second)
 
-        val answer = intent.getIntExtra("number", 0)
+        val answer = intent.getDoubleExtra("number", 0.0)
         textView.text = answer.toString()
 
     }
